Don't use %zu
Some of our bots / compilers don't like it. Wasn't caught by bots because this
is only in debug code.

https://github.com/harfbuzz/harfbuzz/pull/809/commits/825ad378e0f0e6e6bb690894a0d5d25f198be21b

+++ src/hb-ot-cmap-table.hh
@@ -604,7 +604,7 @@ struct cmap
                    + 12 * groups.len; // SequentialMapGroup records
     void *dest = calloc (dest_sz, 1);
     if (unlikely (!dest)) {
-      DEBUG_MSG(SUBSET, nullptr, "Unable to alloc %zu for cmap subset output", dest_sz);
+      DEBUG_MSG(SUBSET, nullptr, "Unable to alloc %lu for cmap subset output", (unsigned long) dest_sz);
       return false;
     }
